Modfies plot_agent_tot_met_div_by_env_wealth_created to move legend to bottom right corner Modifies parse_logs name to be in snake_case Removes redundant code from parse_logs and refactors accordingly Fixes plotting pipeline not able to plot runs that terminate early
