+++ plots/plot_agent_tot_met_div_by_env_wealth_created.plg
@@ -2,7 +2,7 @@ set title "Agent Total Metabolism Divided By Wealth Created"
 set xlabel "Timestep"
 set ylabel "Total Metabolism / Wealth Created"
 set lt 1 lw 2 lc "black"
-set key fixed left top
+set key fixed right bottom
 set xtics nomirror
 set ytics nomirror
 set term pdf
